Split run profile module into focused files
