+++ crates/ironclaw_turns/src/run_profile/mod.rs
@@ -1,0 +1,26 @@
+mod driver;
+mod policy;
+mod refs;
+mod resolver;
+mod snapshot;
+
+pub use driver::{
+    AgentLoopDriver, AgentLoopDriverDescriptor, AgentLoopDriverError, AgentLoopDriverHost,
+    AgentLoopDriverResumeRequest, AgentLoopDriverRunRequest,
+};
+pub use policy::{
+    CancellationPolicy, CheckpointPolicy, PrivilegedRunProfileDimension,
+    RedactedRunProfileProvenance, RedactedRunProfileSource, ResourceBudgetPolicy,
+    RunProfileRequestAuthority, RunProfileResolutionError, RuntimeProfileConstraints,
+    SteeringPolicy,
+};
+pub use refs::{
+    CapabilitySurfaceProfileId, CheckpointSchemaId, ConcurrencyClass, ContextProfileId,
+    LoopDriverId, ModelProfileId, ResourceBudgetTier, RunClassId, RunProfileFingerprint,
+    RunProfileSourceLayer, RunProfileSourceRef, RunnerPoolId, SchedulingClass,
+};
+pub use resolver::{
+    InMemoryRunProfileRegistry, InMemoryRunProfileResolver, RunProfileDefinition,
+    RunProfileResolutionRequest, RunProfileResolver,
+};
+pub use snapshot::ResolvedRunProfile;
