+++ crates/ironclaw_turns/src/run_profile/resolver.rs
@@ -1,306 +1,23 @@
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
-use thiserror::Error;
 
-use crate::{
-    LoopExit, RunProfileId, RunProfileRequest, RunProfileVersion, TurnCheckpointId, TurnId,
-    TurnRunId,
-};
+use crate::{RunProfileId, RunProfileRequest, RunProfileVersion};
 
-macro_rules! profile_ref {
-    ($name:ident, $kind:literal) => {
-        #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-        pub struct $name(String);
-
-        impl $name {
-            pub fn new(value: impl Into<String>) -> Result<Self, String> {
-                let value = value.into();
-                validate_profile_ref($kind, &value)?;
-                Ok(Self(value))
-            }
-
-            #[allow(dead_code)]
-            pub(crate) fn from_trusted_static(value: &'static str) -> Self {
-                debug_assert!(validate_profile_ref($kind, value).is_ok());
-                Self(value.to_string())
-            }
-
-            #[allow(dead_code)]
-            pub(crate) fn from_trusted_string(value: String) -> Self {
-                debug_assert!(validate_profile_ref($kind, &value).is_ok());
-                Self(value)
-            }
-
-            pub fn as_str(&self) -> &str {
-                &self.0
-            }
-        }
-
-        impl Serialize for $name {
-            fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-            where
-                S: serde::Serializer,
-            {
-                serializer.serialize_str(&self.0)
-            }
-        }
-
-        impl<'de> Deserialize<'de> for $name {
-            fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-            where
-                D: serde::Deserializer<'de>,
-            {
-                let value = String::deserialize(deserializer)?;
-                Self::new(value).map_err(serde::de::Error::custom)
-            }
-        }
-    };
-}
-
-profile_ref!(RunClassId, "run_class_id");
-profile_ref!(LoopDriverId, "loop_driver_id");
-profile_ref!(CheckpointSchemaId, "checkpoint_schema_id");
-profile_ref!(ModelProfileId, "model_profile_id");
-profile_ref!(CapabilitySurfaceProfileId, "capability_surface_profile_id");
-profile_ref!(ContextProfileId, "context_profile_id");
-profile_ref!(RunnerPoolId, "runner_pool_id");
-profile_ref!(SchedulingClass, "scheduling_class");
-profile_ref!(ConcurrencyClass, "concurrency_class");
-profile_ref!(ResourceBudgetTier, "resource_budget_tier");
-profile_ref!(RunProfileFingerprint, "run_profile_fingerprint");
-profile_ref!(RunProfileSourceLayer, "run_profile_source_layer");
-profile_ref!(RunProfileSourceRef, "run_profile_source_ref");
-
-fn validate_profile_ref(kind: &'static str, value: &str) -> Result<(), String> {
-    if value.is_empty() {
-        return Err(format!("{kind} must not be empty"));
-    }
-    if value.len() > 128 {
-        return Err(format!("{kind} must be at most 128 bytes"));
-    }
-    if !value
-        .chars()
-        .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-' || c == ':')
-    {
-        return Err(format!(
-            "{kind} must contain only lowercase ASCII letters, digits, _, -, or :"
-        ));
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct AgentLoopDriverDescriptor {
-    pub id: LoopDriverId,
-    pub version: RunProfileVersion,
-    pub checkpoint_schema_id: Option<CheckpointSchemaId>,
-    pub checkpoint_schema_version: Option<RunProfileVersion>,
-}
-
-impl AgentLoopDriverDescriptor {
-    pub fn new(id: impl Into<String>, version: RunProfileVersion) -> Result<Self, String> {
-        Ok(Self {
-            id: LoopDriverId::new(id)?,
-            version,
-            checkpoint_schema_id: None,
-            checkpoint_schema_version: None,
-        })
-    }
-
-    pub fn with_checkpoint_schema(
-        mut self,
-        checkpoint_schema_id: impl Into<String>,
-        checkpoint_schema_version: RunProfileVersion,
-    ) -> Result<Self, String> {
-        self.checkpoint_schema_id = Some(CheckpointSchemaId::new(checkpoint_schema_id)?);
-        self.checkpoint_schema_version = Some(checkpoint_schema_version);
-        Ok(self)
-    }
-}
-
-/// Minimal host marker for the driver contract.
-///
-/// The concrete per-run host facade is owned by the AgentLoopHost work. This
-/// marker keeps this crate's driver contract loop-neutral: drivers receive a
-/// host capability boundary, not raw runtime, provider, process, filesystem,
-/// network, secret, approval, grant, or lease handles.
-pub trait AgentLoopDriverHost: Send + Sync {}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct AgentLoopDriverRunRequest {
-    pub turn_id: TurnId,
-    pub run_id: TurnRunId,
-    pub resolved_run_profile: ResolvedRunProfile,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct AgentLoopDriverResumeRequest {
-    pub turn_id: TurnId,
-    pub run_id: TurnRunId,
-    pub checkpoint_id: TurnCheckpointId,
-    pub resolved_run_profile: ResolvedRunProfile,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Error)]
-pub enum AgentLoopDriverError {
-    #[error("agent loop driver rejected request: {reason}")]
-    InvalidRequest { reason: String },
-    #[error("agent loop driver is unavailable: {reason}")]
-    Unavailable { reason: String },
-    #[error("agent loop driver failed: {reason_kind}")]
-    Failed { reason_kind: String },
-}
-
-/// Userland loop implementation contract.
-///
-/// Implementations own loop mechanics and return a [`LoopExit`] handshake to the
-/// trusted runner. They do not mutate turn state directly and do not receive raw
-/// authority handles.
-#[async_trait]
-pub trait AgentLoopDriver: Send + Sync {
-    fn descriptor(&self) -> AgentLoopDriverDescriptor;
-
-    async fn run(
-        &self,
-        request: AgentLoopDriverRunRequest,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-    ) -> Result<LoopExit, AgentLoopDriverError>;
-
-    async fn resume(
-        &self,
-        request: AgentLoopDriverResumeRequest,
-        host: &(dyn AgentLoopDriverHost + Send + Sync),
-    ) -> Result<LoopExit, AgentLoopDriverError>;
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ResolvedRunProfile {
-    pub run_class_id: RunClassId,
-    pub profile_id: RunProfileId,
-    pub profile_version: RunProfileVersion,
-    pub loop_driver: AgentLoopDriverDescriptor,
-    pub checkpoint_schema_id: CheckpointSchemaId,
-    pub checkpoint_schema_version: RunProfileVersion,
-    pub model_profile_id: ModelProfileId,
-    pub capability_surface_profile_id: CapabilitySurfaceProfileId,
-    pub context_profile_id: ContextProfileId,
-    pub steering_policy: SteeringPolicy,
-    pub cancellation_policy: CancellationPolicy,
-    pub checkpoint_policy: CheckpointPolicy,
-    pub resource_budget_policy: ResourceBudgetPolicy,
-    pub runtime_constraints: RuntimeProfileConstraints,
-    pub runner_pool_id: Option<RunnerPoolId>,
-    pub scheduling_class: SchedulingClass,
-    pub concurrency_class: ConcurrencyClass,
-    pub resolution_fingerprint: RunProfileFingerprint,
-    pub provenance: RedactedRunProfileProvenance,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SteeringPolicy {
-    pub allow_steering: bool,
-    pub allow_interrupt: bool,
-    pub allow_driver_specific_nudges: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CancellationPolicy {
-    pub allow_cancel: bool,
-    pub require_checkpoint_before_cancel: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CheckpointPolicy {
-    pub require_before_model: bool,
-    pub require_before_side_effect: bool,
-    pub require_before_block: bool,
-    pub max_checkpoint_bytes: u64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ResourceBudgetPolicy {
-    pub tier: ResourceBudgetTier,
-    pub max_model_calls: u32,
-    pub max_capability_invocations: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RuntimeProfileConstraints {
-    pub allow_raw_runtime_backend_selection: bool,
-    pub allow_broad_capability_surface: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RedactedRunProfileProvenance {
-    pub sources: Vec<RedactedRunProfileSource>,
-    pub effective_privileges: Vec<PrivilegedRunProfileDimension>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RedactedRunProfileSource {
-    pub layer: RunProfileSourceLayer,
-    pub source_ref: RunProfileSourceRef,
-    pub summary: String,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum RunProfileRequestAuthority {
-    #[default]
-    User,
-    ProductSurface,
-    Admin,
-    System,
-}
-
-impl RunProfileRequestAuthority {
-    fn allows(self, dimension: PrivilegedRunProfileDimension) -> bool {
-        match dimension {
-            PrivilegedRunProfileDimension::LongRunningMission
-            | PrivilegedRunProfileDimension::SpecialDriver
-            | PrivilegedRunProfileDimension::RunnerPool => {
-                matches!(self, Self::ProductSurface | Self::Admin | Self::System)
-            }
-            PrivilegedRunProfileDimension::BroadCapabilitySurface
-            | PrivilegedRunProfileDimension::HighBudget => {
-                matches!(self, Self::Admin | Self::System)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum PrivilegedRunProfileDimension {
-    LongRunningMission,
-    BroadCapabilitySurface,
-    HighBudget,
-    SpecialDriver,
-    RunnerPool,
-}
-
-impl PrivilegedRunProfileDimension {
-    fn category(self) -> &'static str {
-        match self {
-            Self::LongRunningMission => "long_running_mission",
-            Self::BroadCapabilitySurface => "broad_capability_surface",
-            Self::HighBudget => "high_budget",
-            Self::SpecialDriver => "special_driver",
-            Self::RunnerPool => "runner_pool",
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Error)]
-pub enum RunProfileResolutionError {
-    #[error("run profile request is unauthorized for {dimension:?}")]
-    Unauthorized {
-        dimension: PrivilegedRunProfileDimension,
+use super::{
+    driver::AgentLoopDriverDescriptor,
+    policy::{
+        CancellationPolicy, CheckpointPolicy, PrivilegedRunProfileDimension,
+        RedactedRunProfileProvenance, RedactedRunProfileSource, ResourceBudgetPolicy,
+        RunProfileRequestAuthority, RunProfileResolutionError, RuntimeProfileConstraints,
+        SteeringPolicy,
     },
-    #[error("run profile is unavailable: {profile_id}")]
-    ProfileUnavailable { profile_id: String },
-    #[error("invalid run profile request: {reason}")]
-    InvalidRequest { reason: String },
-}
+    refs::{
+        CapabilitySurfaceProfileId, CheckpointSchemaId, ConcurrencyClass, ContextProfileId,
+        LoopDriverId, ModelProfileId, ResourceBudgetTier, RunClassId, RunProfileFingerprint,
+        RunProfileSourceLayer, RunProfileSourceRef, RunnerPoolId, SchedulingClass,
+    },
+    snapshot::ResolvedRunProfile,
+};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RunProfileResolutionRequest {
